Move to test directory (where it belongs) -- also will let windows build glooctl again
