+++ test/kube2e/knative/knative_suite_test.go
@@ -18,7 +18,6 @@ import (
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"github.com/solo-io/gloo/pkg/cliutil"
 	skhelpers "github.com/solo-io/solo-kit/test/helpers"
 )
 
@@ -26,7 +25,7 @@ func TestKnative(t *testing.T) {
 	if testutils.AreTestsDisabled() {
 		return
 	}
-	cliutil.RegisterGlooDebugLogPrintHandlerAndClearLogs()
+	helpers.RegisterGlooDebugLogPrintHandlerAndClearLogs()
 	skhelpers.RegisterCommonFailHandlers()
 	skhelpers.SetupLog()
 	RunSpecs(t, "Knative Suite")
